Fixes for some mem-leaks: most changes where pretty basic (i.e. adding deletes).
In some cases affecting AliESDEvent and AliTPCtrackerMI memory leaks introduced by placement-new were fixed. The placement-new sometimes overwrote pointers to memory allocated by the class itself.The placement-new was replaced with assigments where it was considered appropiate (for this some operator= were modified) or an explicit call to the destructor before was used.

Christian