AliMUONRecoParam.cxx AliMUONRecoParam.h
Adding new methods to :

- be able to Bypass station 4 or station 5 or both (as before) (or
none ;-) )

- get/set limits on calibration parameters (pedestal,gain,HV)

- get/set the "pad goodness mask", which is the basis to tell whether
a pad is to be rejected or not.

Note that this commit only adds those methods. They are not yet used
anywhere...

AliMUONPadStatusMaker.cxx
AliMUONPadStatusMaker.h

Adding one piece of information to the pad status = the
occupancy. That occupancy is fetched from the
MUON.QA.RunNumber.Cycle.root file, if that file exists and if is
contains the relevant AliMUONVTrackerData object...