The line
   if (edep<=0) continue;

in AliFMDBaseDigitizer::DigitizeHits was introduced - wrongly - by 
Peter Hristov.  It _will_ break the digitisation and the following 
reconstruction.  

The behviour of the digitisation models exactly the front-end as it 
should (no matter what memory consumption it may entail).  The check 
should be on zero suppression, since that's what models the front-end - 
if zero suppression is turned on in the front-end, then we can suppress 
empty digits - otherwise we shoud never do that.  Note, that the line 
affects _both_ normal digitisation and digitisation for summable digits, 
since the condition is on the energy deposition and not on the actual 
number of counts.  If this line should go anywhere, it should be in 
the possible overloaded AliFMDSDigitizer::AddDigit - not here.

Perhaps this change should be propogated to the release ASAP?