new function PlotTrackletsVsFindable (Markus)
Inside the function
- added a documentation
- Changed the way in which I am doing the Kalman filtering (first step
inwards, the fit outwards in order to do extrapolation, done on a local
copy of the track)
- enabled the debug stream

In Functions PostProcess and GetRefFigures  

- enabled the PostProcessing

Further on I made the TRD geometry a class member.

I also made a change in the runscript that from now on loads the
GeoManager (From the database, correct database has to be initialized).