AliMUONDigitCalibrator
Adding a new calibration option : INJECTIONGAIN, where we use capacitances and gain, but with the 
 "injection" gain, aka Emelec gain, i.e. the gain from the factory. 
At the same time, make sure that whatever the calibration option chosen, the computed charge is always in the same unit, mV.
Before, the charge was in ADC when no gain was applied, and in fC^-1 (?) otherwise.  

Other changes : 

Just a propagation of the introduction of this new option, mainly in mchview-related classes.

Alberto B. and Jean-Luc C.