bugfix https://savannah.cern.ch/bugs/?55067 Updating AliHLTMessage to TMessage r27689
HLT encountered a problem when streaming the AliESDEvent. The problem occurs if the
TClonesArray members contain more then one entry. In this case, members of the individual
objects have been mixed.

The problem turned out to be solely on the serialization part, and did not effect the
restoration part. This has been proven by using either TMessage and AliHLTMessage on the
sender, and only AliHLTMessage on the receiver side. Merging the recent changes in
TMessage solved the issue.