implementing the short info option which includes a TNames array instead of input objects and trigger decisions adding the option -include-input=both to include both the objects and short info bugfix in the parsing of the -include-input option
There is currently a problem in the streaming of the arrays of the GlobalTriggerDecision which causes the
short info not to show up in the extracted object when there are no trigger decisions. Furthermore only the
the fisrt object of the input object array is there after extracting (in mode 'both'). The second one which
is the CTPData object is missing. Under investigation.