Major upgrade to the package, in order to speed-up the execution and remove some unnecessary features which complicate too much the preparation of configuration. Added some new cuts used for current analysis, and some monitor analysis task in order to check some quality cuts to tune them properly.
AliRsnDaughter:
- added a TLorentzVector datamember which will contain all kinematic informations
  (one for default input and another for MC if necessary)
  >> removed most of the kinematic-like getters
  >> access to kinematics of the track through the getter which returns this object
  >> eases a lot the procedure of summing up without errors
  NOTE: now the AliRsnDaughter should be able to deal also with V0s, in order to
        search for *charged* resonances (this is under test, but it is basically implemented)

AliRsnPairParticle --> renamed into AliRsnMother
- added two TLorentzVector which are set from two AliRsnDaughters and two masses
  assigned to them


The AliRsnCutManager (renamed class) has been reviewes as follows:
- removed the slot related to cuts on events, because it is redundant
  if repeated on all AliRsnPair objects, it has been moved at the level
  of the AnalysisManager
- added the possibility to specify some cuts only for the first or second daughter
  so there are three different slots: common (checked for both), first and second (checked for one only)

AliRsnFunctionAxis was renamed as AliRsnValue for meaning reasons.

AliRsn Analysis tasks will contain only one analysis manager,
since the use of them in the train (which should be the default)
stores all outputs into a single file, so there is no reason to have
different analysis manager, which cause confusion when configuring the task.

AliRsnPair becam a virtual base classe for different output systems
the 'old' version which stores histograms is now AliRsnPairFunctions, and works like before.
A new possibility was added to save output as a TNtuple, with the class AliRsnPairNtuple,
which needs to be initialized just defining what values it has to store for each entry, 
which are defined using the same class AliRsnValue which is used to create the AliRsnFunctions
which are used as the output for the previous version of this class.
There is the possibility to implement new output formats using this base class.

REMOVED CLASSES:
- AliRsnVManager (not needed any longer)
- AliRsnHistoDef (obsolete, not used since long time)
- AliRsnPairManager (unnecessary step more in configuration)
- AliRsnPIDDefESD (replaced by a dedicated AliRsnCut for PID)
- AliRsnAnalysisTrackEffSE (never used)
- AliRsnCutESDTrigger (better to use AliPhysicsSelection)