Start translation of the macros helper.C and AliTRDperformanceTrain.h into a class AliTRDPWG1helper. (Markus)
Reason is the following: Whenever I want to run
the train with a different configuration (like now with the new trigger)
and I copy the add macros to a test directory, rootcint doesn't find the
two macros unless I copy them to the same path. This is something I
don't like because I have to do the same on the grid. It is better to
have this compiled inside a class (as static members) so one only has to
load the library PWG1 and the simply use it.