===============================================  Updates to the correction class ===============================================
1) Separated statistical and systematic uncertainties (added comments on the classes)
a) Statistical uncertainties are kept on the histograms
b) Systematic uncertainties (but feed-down related) stored on a TGraphAsymmErrors
c) Feed-down systematic uncertainties are kept in other TGraphAsymmErrors. The theoretical extreme and conservative approach are both calculated. 

2) Uncertainties from:
a)  the acceptance x efficiency corrections are taken from the efficiency histograms and propagated to the final cross-section and feed-down-Nb uncertainties (it is now missing, on purpose, on the feed-down-fc uncertainties)
b) signal counting systematic uncertainties can be fed externally via a TGraphAsymmErrors, they will be propagated to the systematic uncertainties (not related to feed-down, point 1.b.)

3) Bug fixes:
a) related to naming & cloning histograms to evict problems of object ownership 
b) on the reconstructed spectra loop (Yifei) 

4) Added D* on the example macro 

===============================================