AliAnaParticleHadronCorrelation: Remove not necessary requests for isolated AODs, change TH1 histograms array for different pt trigger bins to TH2 with pt of trigger in x axis and now array is for pt associated bins. And cosmetics
AliAnaChargedParticles: Remove TH3D histograms, transform into TH2F, remove Vertex TH3 histogram. Do not get the charge of the track but set it when filling the AODs in the AOD and get if from there when filling the histograms. And cosmetics

AliAnaPhoton: Produce track-cluster matching correlation histograms when the track matching rejection is requested and not when also the shower shape PID is on

Reader: Delete the AliESDtrackCuts pointer when it is going to be set externally