Modifications to plot thermal fits vs data comparisons + cosmetics
Added functions to plot results in GSI or THERMUS format
Changed way labels are plotted (both for ratios and yields). Now they
are added by hand at the top of the macro.
Added markers for extrapolated and not-in-fit particles.
Improved GetHistoyields
Added possibility to overwrite defualt shift of histograms
Added shorthands for the different yields vs thermal plots