Add macros to populate a local OCDB
A global steering macro calls a macro per detector, which in turn calls
the detector specific macros to generate the CDB objects for the given
detector. These objects are as those which were kept in the OCDB of the
repository, that is general-purpuse default (0-inf run-range) CDB
objects, whose purpose is to allow running local test (as those in
test/) without needing an AliEn connection.

Make ACORDE AD EMCAL GRP ZDC objects
Fix the macro to make LHCData