Changing precision of huffman weight member from float to double
float precission for the weight leads to a non-optimal huffman table if a
large sample is used and float precision is not enough to represent small
differences in large occurrence of values. Precision has been changed to
double internally. The new double precision weight member is however defined
transient and the float precision is kept for storage in order to avoid
increase of the (already big) huffman table. The float precision is approached
by scaling weights of all nodes by the exponent of the least non-zero weight.
This is not entirely accurate but given the informative character of the weight
member after the generation of the huffman codes, this is accepted.