Improve performance of AliSimDigits::ExpandTrackBuffer1()
AliSimDigits::ExpandTrackBuffer1() seems to be responsible for roughly
20% of the time in ppbench and 11% in gun reconstruction (albeit it's as
low as 0.25% on PbPbbench, where of course tracking takes over).

The reason is that it exploits the fact that TArrayI currently sets the
memory to zero, via memset. This changes things to use calloc, which
also returns zero-ed memory but exploits kernel features as zero pages
to avoid as much zeroing as possible.