Turn on -fvisibility-inlines-hidden option
This option will make sure that all functions which are inlined will not
also generate an out-of-line version of the same code, in case they need
to be called explicitly from dlopen + dlsym. In my limited PbPbbench
tests this shows 20MB or so reduction between code and link time tables.