Implement new mapping procedure for DCAL
The mapping procedure needed to be changed for DCAL. As the number of
DDLs is larger than 32, the old mapping procedure using bits to identify
DDLs doesn't work any more. Instead we expect as specification the DDL
ID relative to the EMCAL offset. Only for the raw analyzer the specification
is still ported as it is needed in the digit maker - in all other components
it is set to 0. As the mapping procedure differs now for EMCAL and PHOS,
several components diverge.

Furthermore components which by now expected to run on SINGLE DDLs are
modified to handle also multiple DDLs. This is the case for
- RAW Analyzer - EMCAL only
- TRU Analyzer
- Digit Maker - EMCAL only
PHOS components are not touched by the changes.