Bugfix for retrieving single bits from AliHLTDataInflaterHuffman
The mehod NextValue of AliHLTDataInflaterHuffman buffers the bits which have
been already read from the stream but not yet used in the Huffman decoding
of the lat symbol. The read-ahead approach is needed as code length is unknown
before the decoding.

Method InputBit was so far reading directly from the stream at a potentially
wrong position position. Fixed by overloading InputBit and reading from the
buffered bit field first.

There is still possible dislocation of the read pointer when using NextValue
and InputBits interleaved from the caller. A possible solution is to implement
the caching in the base class.