Introducing modes 'global' and 'local' for extraction of TPC padrow
Default is 'global' and corresponds to the behavior until now.
Compressed clusters are stored with the local padrow number within the
readout partition. To get the global padrow number in mode 'global, the
decoder adds the offset of the first padrow in the partition. In mode
'local', padrow number is within partition.