Bugfix: rewinding of bit position of the read pointer
Due to a local variable which is naturally discarded at the end of the method,
rewinding of position within the same byte was not working. This resulted in
incorrect read position when a number of bits was directly read from the inflater.
This case happens in the unpacking of track model clusters when reading the
number of clusters stored on one padrow.