Simple shared pointer implementation for ALICE
Usage of c++11 code is forbidden in AliRoot untill further
notice. However shared pointers are a powerfull tool one
wants to have awailable. As shared pointers are part of
the standard library, an implementation can be found backporting
the functionality of a shared pointer to AliRoot.

N.B.: shared pointers are by definition template classes - no
dictionaries are created. Consequently they cannot be used in
classes unless added in specific casses to the LinkDef file +
specifications in the source files.